backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
ultralytics==8.2.2
easyocr==1.7.1
opencv-python==4.9.0.80
numpy==1.26.4
python-dotenv==1.0.1
aiofiles==23.2.1
